requirements.txt
astroid==3.2.2
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
dill==0.3.8
filelock==3.14.0
Flask==3.0.3
fsspec==2024.6.0
huggingface-hub==0.23.3
idna==3.7
isort==5.13.2
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==2.1.5
mccabe==0.7.0
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
packaging==24.0
parso==0.8.4
pillow==10.3.0
platformdirs==4.2.2
PyJWT==2.8.0
pylint==3.2.3
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.13.1
sentence-transformers==3.0.1
sympy==1.12.1
threadpoolctl==3.5.0
tokenizers==0.19.1
tomlkit==0.12.5
torch==2.3.1
tqdm==4.66.4
transformers==4.41.2
typing_extensions==4.12.2
urllib3==2.2.1
Werkzeug==3.0.3
